tests/wpt/mozilla/tests/mozilla/bluetooth/requestDevice/acceptAllDevices/device-with-empty-name.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.empty_name_heart_rate);
    return window.navigator.bluetooth.requestDevice({acceptAllDevices: true})
    .then(device => assert_equals(device.name, ''));
}, 'Device with empty name and no UUIDs nearby. Should be found if acceptAllDevices is true.');
</script>
